add logger in transformer, move helper functions out of predictor.go
